requirements.txt
asgiref             3.7.2
Django              4.2.2
django-cors-headers 4.0.0
djangorestframework 3.14.0
pip                 22.3.1
psycopg2            2.9.6
pytz                2023.3
setuptools          65.5.0
sqlparse            0.4.4
typing_extensions   4.6.3
tzdata              2023.3